refactor(web): reorganize page files by user and org
